feat: PRP-003 PostgreSQL database foundation + Pre-commit hooks
Implements PRP-003: PostgreSQL Database Foundation with comprehensive
testing, migrations, and quality automation.

Database Infrastructure:
- SQLAlchemy async engine with connection pooling (pool_size=10, max_overflow=20)
- Alembic migrations for database versioning
- Support for PostgreSQL (asyncpg) and SQLite (aiosqlite) for testing

Models Created:
- User: Telegram user info with friend status, language support
- Message: Linear chat history for RAG (user_id, chat_id, text, language, timestamp)
- Fact: User facts storage for personalization
- Stat: User statistics and metrics tracking

Testing:
- 11 new tests (9 unit + 2 E2E) - All passing ✅
- Total: 20 tests passing in 1.76s
- E2E tests cover message storage/retrieval flows
- Unit tests cover all CRUD operations

Quality Automation:
- Pre-commit hooks configured (.pre-commit-config.yaml)
- Automated: ruff linting/formatting, mypy, file checks, pytest
- CONTRIBUTING.md enhanced with pre-commit workflow
- All quality checks passing before commit

Additional:
- PRP-012: Analytics & Observability framework designed
- Research completed: LangSmith, Prometheus, Grafana
- Privacy-first approach with GDPR compliance requirements

Definition of Done:
✅ PostgreSQL database setup and connected
✅ SQLAlchemy models created (User, Message, Fact, Stat)
✅ Alembic migrations initialized and working
✅ Linear message history storage implemented
✅ Connection pooling configured
✅ Unit tests for database models
✅ E2E test for message storage and retrieval

Implements: PRP-003
Related: PRP-012 (Analytics framework designed)

+++ models/message.py
@@ -1,7 +1,7 @@
 """Message model for dcmaidbot."""
 
 from datetime import datetime
-from sqlalchemy import BigInteger, String, Text, DateTime, ForeignKey
+from sqlalchemy import BigInteger, String, Text, DateTime, ForeignKey, Integer
 from sqlalchemy.orm import Mapped, mapped_column
 from database import Base
 
@@ -11,9 +11,9 @@ class Message(Base):
 
     __tablename__ = "messages"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     user_id: Mapped[int] = mapped_column(
-        BigInteger, ForeignKey("users.id"), nullable=False
+        Integer, ForeignKey("users.id"), nullable=False
     )
     chat_id: Mapped[int] = mapped_column(BigInteger, nullable=False, index=True)
     message_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
