+++ models/user.py
@@ -1,7 +1,7 @@
 """User model for dcmaidbot."""
 
 from datetime import datetime
-from sqlalchemy import BigInteger, String, Boolean, DateTime
+from sqlalchemy import BigInteger, String, Boolean, DateTime, Integer
 from sqlalchemy.orm import Mapped, mapped_column
 from database import Base
 
@@ -11,7 +11,7 @@ class User(Base):
 
     __tablename__ = "users"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     telegram_id: Mapped[int] = mapped_column(
         BigInteger, unique=True, nullable=False, index=True
     )
